requirements.txt
flask
gunicorn
neo4j
requests
python-dotenv
rdkit
openai
numpy
reportlab
